Moving all components outside of app dir, updating references
