+++ src/components/footer/footertypes.tsx
